requirements.txt
# Видео и изображения
opencv-python-headless>=4.8.0
Pillow>=10.0.0

# Модель для поиска по фото/тексту (CLIP)
torch>=2.0.0
transformers>=4.30.0
ftfy

# Веб-сервер
flask>=3.0.0
flask-cors>=4.0.0

# Утилиты
numpy>=1.24.0
tqdm>=4.65.0
